requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
asttokens=2.4.1=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
bzip2=1.0.8=hd590300_5
ca-certificates=2024.2.2=hbcca054_0
certifi=2022.12.7=pypi_0
charset-normalizer=3.1.0=pypi_0
cmake=3.26.3=pypi_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.2.0=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.8.1=py39h3d6467e_0
decorator=5.1.1=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
filelock=3.11.0=pypi_0
fonttools=4.47.0=pypi_0
huggingface-hub=0.13.4=pypi_0
idna=3.4=pypi_0
importlib-metadata=7.1.0=pyha770c72_0
importlib-resources=6.1.1=pypi_0
importlib_metadata=7.1.0=hd8ed1ab_0
ipykernel=6.29.3=pyhd33586a_0
ipython=8.12.0=pyh41d4057_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=pypi_0
jupyter_client=8.6.1=pyhd8ed1ab_0
jupyter_core=5.7.2=py39hf3d152e_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.40=h41732ed_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=13.2.0=h807b86a_5
libgomp=13.2.0=h807b86a_5
libnsl=2.0.1=hd590300_0
libsodium=1.0.18=h36c2ea0_1
libsqlite=3.45.2=h2797004_0
libstdcxx-ng=13.2.0=h7e041cc_5
libuuid=2.38.1=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.2.13=hd590300_5
lit=16.0.1=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.8.2=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mpmath=1.3.0=pypi_0
ncurses=6.4.20240210=h59595ed_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.1=pypi_0
numpy=1.24.2=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
openssl=3.2.1=hd590300_1
packaging=23.1=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.1.0=pypi_0
pip=24.0=pyhd8ed1ab_0
platformdirs=4.2.0=pyhd8ed1ab_0
prompt-toolkit=3.0.42=pyha770c72_0
prompt_toolkit=3.0.42=hd8ed1ab_0
psutil=5.9.8=py39hd1e30aa_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pygments=2.17.2=pyhd8ed1ab_0
pyparsing=3.1.1=pypi_0
python=3.9.19=h0755675_0_cpython
python-dateutil=2.8.2=pypi_0
python_abi=3.9=4_cp39
pyyaml=6.0=pypi_0
pyzmq=25.1.2=py39h8c080ef_0
readline=8.2=h8228510_1
regex=2023.3.23=pypi_0
requests=2.28.2=pypi_0
setuptools=69.2.0=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.11.1=pypi_0
tk=8.6.13=noxft_h4845f30_101
tokenizers=0.13.3=pypi_0
torch=2.0.0=pypi_0
tornado=6.4=py39hd1e30aa_0
tqdm=4.65.0=pypi_0
traitlets=5.14.2=pyhd8ed1ab_0
transformers=4.28.1=pypi_0
triton=2.0.0=pypi_0
typing-extensions=4.5.0=pypi_0
typing_extensions=4.10.0=pyha770c72_0
tzdata=2024a=h0c530f3_0
urllib3=1.26.15=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.43.0=pyhd8ed1ab_0
xz=5.2.10=h5eee18b_1
zeromq=4.3.5=h59595ed_1
zipp=3.17.0=pyhd8ed1ab_0
